Operators file divided into smaller files, because it started deliveritg too much data
