+++ src/data/js/operators.js
@@ -1,3 +1,8 @@
+import getComparationOperatorsData from './operators_files/comparasionOperators';
+import getAssignmentOperatorsData from './operators_files/assignmentOperators';
+import getArithmeticsOperatorsData from './operators_files/arithmeticOperators';
+
+
 let data =     {
     summary: 'operators',
     title: 'operators',
@@ -33,1221 +38,22 @@ let data =     {
 
 
 
-                {
-                    elementType:'Headline-3',
-                    content:'Comparasion operators'
-                },
-                {
-                    elementType: 'Paragraph',
-                    content: `
-                        Comparation operators evaluate the left and the right operand expression, compares them and 
-                        returns the result. Operands of the non strict comparasions (<code>==</code>, <code>!=</code>) 
-                        that are not of the same type are converted to the same type, usually numeric. 
-                    `
-                },
-                {
-                    elementType:'SmallHeadline',
-                    content:'Avoid non strict comparations: <code>==</code>, <code>!=</code>'
-                },
-                {
-                    elementType: 'Paragraph',
-                    content: `
-                        The algorithm:
-                    `
-                },
-                {
-                    elementType: 'UnsignedList',
-                    content: [
-                        `If both operands are of the same type, compare with strict algorithm`,
-                        `Undefined and null are considerated equal,`,
-                        `If string and number, convert to string,`,
-                        `If nonboolean and boolean, convert boolean to number and start again,`,
-                        `Id an object is compared with a number or a string, convert the object to a primitive`,
-                        `If both are strings, then this is already covered, operands are of the same type and 
-                        are compared with a strict algorithm`,
-                        `If none of above then operands are not equal`
-                    ]
-                },
+                ...getComparationOperatorsData(),
 
-                {
-                    elementType: 'Paragraph',
-                    content: `
-                        Problems:
-                    `
-                },
-                {
-                    elementType: 'UnsignedList',
-                    content: [
-                        `<code>2 == true is false</code> as true is converted to 1,`,
-                        `<code>'2' == true is false</code>`,
-                        `<code>'\n\t123\r ' == 123 is true</code> and this may not be ok`,
-                        `<code>{} == '[object Object]' is true</code>`,
-                        `<code>[] == 0 is true</code>`,
-                        `<code>['12'] == 12 is true</code>`,
-                        `<code>new String('') == new String('') is false</code>`,
-                        `<code>NaN == NaN is false</code>, isNaN() should be used`
-                    ]
-                },
-                {
-                    elementType: 'Paragraph',
-                    content: `
-                        Instead of non strict comparations use:
-                    `
-                },
-                {
-                    elementType: 'Paragraph',
-                    content: `
-                    <table class = "table">
-                    <thead><tr><th>Do not use</th><th>Use</th></tr></thead>
-                    <tbody>
-                        <tr>
-                            <td><code>if (x != null)</code></td>
-                            <td><code>if (x)</code> or <code>if (x !== undefined && x != null)</code></td>
-                        </tr>
-                        <tr>
-                            <td><code>if (x == 123)</code></td>
-                            <td><code>if (Number(x) === 123)</code></td>
-                        </tr>
-                        <tr>
-                            <td><code>isNaN == isNaN</code></td>
-                            <td><code>isNaN()</code></td>
-                        </tr>
-                    </tbody>
-                    </table>
-                    `
-                },
 
-                {
-                    elementType: 'ConditionalArray',
-                    headItems: [
-                        'Operator','Name','Usage','Description'
-                    ],
-                    content: [
-                        {
-                            [Symbol('title')]:'==',
-                            [Symbol('code')]:`
-                            
-<pre>
-null == undefined; // true
-3 == true; // false; as boolean is converted to a number,
-// true is 1
-1 == true; // true;
-0 == false; // true;
 
-'' == false; // true;
-'1' == true; // true;
-'2' == true; // false, as true is converted to 1,
-'some string' == true // false, as true is converted to 1,
-({} == '[object Object]') // true, as object is converted to a 
-// primitive using a <b>@@toStringTag</b> well known symbol
-</pre>                            
-                            
-                            `,
-                            Operator: '<code>==</code>',
-                            Name: `Equal`,
-                            Usage: 'g() == f()',
-                            Description: `true if both opernads are equal, if operands are of a different type try to
-                            convert them to the same type: the number type is preferred, objects are converted to 
-                            primitives. <strong>Use strict operator in stead</strict>`
-                        },
+                ...getArithmeticsOperatorsData(),
 
 
 
 
 
-                        {
-                            [Symbol('title')]:'!=',
-                            [Symbol('code')]:`
-                            
-<pre>
-null != undefined; // false
-3 != true; // true; as boolean is converted to a number,
-// true is 1
-1 != true; // false;
-0 != false; // false;
+                ...getAssignmentOperatorsData(),
 
-'' != false; // false;
-'1' != true; // false;
-'2' != true; // true, as true is converted to 1,
-'some string' != true // true, as true is converted to 1,
-({} != '[object Object]') // false, as object is converted to a 
-// primitive using a <b>@@toStringTag</b> well known symbol
-</pre>                            
-                            
-                            `,
-                            Operator: '<code>!=</code>',
-                            Name: `Equal`,
-                            Usage: 'g() == f()',
-                            Description: `true if both opernads are not equal, if operands are of a different type try to
-                            convert them to the same type: the number type is preferred, objects are converted to 
-                            primitives. <strong>Use strict operator in stead</strict>`
-                        },
 
 
+        
 
-                        {
-                            [Symbol('title')]:'===',
-                            [Symbol('code')]:`
-                            
-<pre>
-NaN === NaN; // false !!
-// isNaN() should be used
-
-+0 === -0; // true;
-
-2 === 2; // true;
-2 === 1; // false;
-2 === '2' // false;
-'some caption' === 'some caption' // true;
-'some caption' === 'some caption ' // false; additional space at the end
-null === undefined // false
-
-let obj1 = {a:1, b:2};
-let obj2 = {a:1, b:2};
-let obj3 = obj1;
-obj1 === obj2; // false; the keys and values are the same,
-// the type is the same, but obj1 and obj2 refere to different
-// places in memory, these objects are not the same
-obj1 === obj3; // true; as obj3 was assigned the value obj1 is pointing to
-// so obj1 and obj3 are the same objects
-</pre>                            
-                            
-                            `,
-                            Operator: '<code>===</code>',
-                            Name: `Strict equal`,
-                            Usage: 'g() === f()',
-                            Description: `Returns true, if both operands are of the same type and have the same value.
-                            If objects compared, then it is compared if both variables point to the same object.`
-                        },
-
-
-
-                        {
-                            [Symbol('title')]:'!==',
-                            [Symbol('code')]:`
-                            
-<pre>
-2 !== 2; // false;
-2 !== 1; // true;
-2 !== '2' // true;
-'some caption' !== 'some caption' // false;
-'some caption' !== 'some caption ' // true; additional space at the end
-null !== undefined // true
-
-let obj1 = {a:1, b:2};
-let obj2 = {a:1, b:2};
-let obj3 = obj1;
-obj1 !== obj2; // true; the keys and values are the same,
-// the type is the same, but obj1 and obj2 refere to different
-// places in memory, these objects are not the same
-obj1 !== obj3; // false; as obj3 was assigned the value obj1 is pointing to
-// so obj1 and obj3 are the same objects
-</pre>                            
-                    
-                            `,
-                            Operator: '<code>!==</code>',
-                            Name: `Strict equal`,
-                            Usage: 'g() !== f()',
-                            Description: `Returns true, if both operands are not of the same type or have different value.
-                            If objects compared, then it is compared if both variables point to the same object.`
-                        },
-
-
-
-
-                        {
-                            [Symbol('title')]:'>',
-                            [Symbol('code')]:`
-                        <ul>
-                            <li>Convert objects to primitives</li>
-                            <li>Strings should be compared with the lexicography comparation of 16-bit code units</li>
-                            <li>Try to convert both operands to numbers</li>
-                            <li>Symbols cannot be compared, as they are not convertable to a number</li>
-                        </ul>
-                                                    
-<pre>
-
-3 > 2; //true;
-[3] > [2]; // true;
-({a:2}>{a:1}); // false
-({a:2}<{a:1}); // false
-'3' > 2; // true
-'2' > 3; // false
-'a' > 'b'; // false;
-'b' > 'a'; // true;
-'bcd' > 'abc'; // true;
-'b' > '1'; // true;
-'b' > 1; // false as b is not convertable to a number
-'b' < 1; // false as b is not convertable to a number
-
-BigInts
-3n > 2n; // true;
-3n > 2; // true;
-3 > 2n; // true;
-
-Undefined
-undefined > 3; // false;
-3 > undefined; // false;
-
-Null
-3 > null; // true
-true > null; // true;
-null < 'some str'; // false as not comparable
-null > 'some str'; // false
-
-Boolean
-true > false; // true
-true > 0; // true
-true > '0'; // true
-true > 'a'; // false
-true < 'a'; // false
-
-NaN
-NaN > 3; // false
-3 > NaN; // false
-</pre>                            
-                    
-                            `,
-                            Operator: '<code>></code>',
-                            Name: `Greater than`,
-                            Usage: 'g() > f()',
-                            Description: `Returns true, left operand is greater than the right one.`
-                        },
-
-
-
-
-                        {
-                            [Symbol('title')]:'<',
-                            [Symbol('code')]:`
-                        <ul>
-                            <li>Convert objects to primitives</li>
-                            <li>Strings should be compared with the lexicography comparation of 16-bit code units</li>
-                            <li>Try to convert both operands to numbers</li>
-                            <li>Symbols cannot be compared, as they are not convertable to a number</li>
-                        </ul>
-                                                    
-<pre>
-2 < 3; //true;
-[1] < [2]; // true;
-({a:2}<{a:1}); // false
-({a:2}>{a:1}); // false
-'1' < 2; // true
-'2' < 1; // false
-'a' < 'b'; // false;
-'b' < 'a'; // true;
-'bcd' < 'xyz'; // true;
-'2' < 'd'; // true;
-'b' < 1; // false as b is not convertable to a number
-'b' > 1; // false as b is not convertable to a number
-
-BigInts
-1n < 2n; // true;
-1n < 2; // true;
-1 < 2n; // true;
-
-Undefined
-undefined < 3; // false;
-3 < undefined; // false;
-
-Null
-null < 3; // true
-null < true; // true;
-null < 'some str'; // false as not comparable
-null > 'some str'; // false
-
-Boolean
-false < true; // true
- 0 < true;   // true
-'0' < true;  // true
-'a' < true;  // false
-'a' > true;  // false
-
-NaN
-NaN < 3; // false
-3 > NaN; // false
-</pre>                            
-                    
-                            `,
-                            Operator: '<code><</code>',
-                            Name: `Less than`,
-                            Usage: 'g() < f()',
-                            Description: `Returns true, left operand is smaller than the right one.`
-                        },
-
-
-
-                        {
-                            [Symbol('title')]:'<=',
-                            [Symbol('code')]:`
-                        <ul>
-                            <li>Convert objects to primitives</li>
-                            <li>Strings should be compared with the lexicography comparation of 16-bit code units</li>
-                            <li>Try to convert both operands to numbers</li>
-                            <li>Symbols cannot be compared, as they are not convertable to a number</li>
-                        </ul>
-                                                    
-<pre>
-Numbers
-1 <= 2; // true;
-3 <= 3; // true;
-
-BigInt
-1n <= 2n; // true;
-3n <= 3n; // true;
-1n <= 2; // true;
-3n <= 3; // true;
-
-Strings
-'5n' <= 6; // false as '5n' is not converted into BigInt
-'abc' <= 'cde'; // true;
-'abc' <= 'abc'; // true;
-'abc' <= '0bc'; // false;
-'4' <= 4; // true;
-'a' <= 4; // false, not comparable
-'a' >= 4; // false;
-
-Symbols
-let s1 = Symbol();
-let s2 = Symbol();
-s1 <= s2; //Error, symbols cannot be converted to a number,
-// moreover symbols are unique, but there is no possiblity to get a symbol
-// real value, so no point to compare them
-
-undefined and null
-null >= undefined; // false;
-null <= undefined; // false;
-null <= true; // true;
-true <= null; // false
-null <= 0; // true
-0 <= null; // true
-undefined <= 0; // false
-0 <= undefined; // false
-
-Booleans
-true <= false; // false;
-false <= true; // true;
-true <= 1; // true;
-true <= 4; // true;
-true <= 's'; // false
-'s' <= true; // false
-null <= false; // true;
-1n <= true; // true;
-0n <= true; // true;
-0n <= false; // true;
-1n <= false; // false;
-
-
-Objects
-({} <= {}); // TRUE
-({a:2}<={a:1}); // TRUE
-({a:1, b:2, c:3} <= {a:0}) // TRUE
-({a:1, b:2, c:3} >= {a:0}) // TRUE
-({a:1, b:2, c:3} <= []); // false
-({a:1, b:2, c:3} >= []); // true
-({a:1, b:2, c:3} <= [1, 2, 3, 4, 5]); // false
-({a:1, b:2, c:3} >= [1, 2, 3, 4, 5]); // true
-
-Arrays
-[2,3] <= [1,2,3]; // false
-[2,3] >= [1,2,3]; // true
-[5] <= [5,6]; // true
-[5] >= [5,6]; // true
-// seems the first value is compared
-</pre>                            
-                    
-                            `,
-                            Operator: '<code><=</code>',
-                            Name: `Less than`,
-                            Usage: 'g() <= f()',
-                            Description: `Returns true, left operand is smaller than or equal to the right one.`
-                        },
-
-
-
-                        {
-                            [Symbol('title')]:'>=',
-                            [Symbol('code')]:`
-                       Please refer to the less or equal than <code><=</code> operator examples
-                                                    
-<pre>
-</pre>                            
-                    
-                            `,
-                            Operator: '<code>>=</code>',
-                            Name: `Less than`,
-                            Usage: 'g() >= f()',
-                            Description: `Returns true, left operand is greater than or equal to the right one.`
-                        },
-
-
-
-
-
-
-
-
-
-
-
-            ]
-        },
-
-
-
-
-
-
-
-        {
-            elementType:'Headline-3',
-            content:'Arithmetic operators'
-        },
-        {
-            elementType: 'Paragraph',
-            content: `
-                Arithmetic operators take left and right operand, perform an opperation and return the result. 
-                In case types of operands do not match, an attempt to convert operands to a number is being performed.
-                BigInt type will not be converted to a number by default, though <code>Number(2n)</code> works.
-                In general if there is no possiblity to convert not matching type to a number, there will be an error.
-                There is an exception, the addition operator <code>+</code> is capable of joining strings. Moreover
-                this operator will convert numbers to strings, so <code>5 + '5' === '55'</code>
-            `
-        },
-
-        {
-            elementType: 'ConditionalArray',
-            headItems: [
-                'Operator','Name','Usage','Description'
-            ],
-            content: [
-                {
-                    [Symbol('title')]:'+',
-                    [Symbol('code')]:`
-                    <ul>
-                        <li>Both numeric operands: returns a sum</li>
-                        <li>At least one operand is a string: converts second operand to string and reutrns a string</li>
-                        <li>undefined + number is NaN</li>
-                        <li>undefined + null is NaN</li>
-                        <li>null + number is the number</li>
-                        <li>In case of an object (and Array) and any other operand converts to a string, where 
-                        object conversion is '[object Object]'</li>
-                        <li>A BigInt will not be automaticly converted into a number, but will be converted to 
-                        a string if the second operand is a string</li>
-                    </ul>                   
-<pre>
-4 + 5 // 9
-4 + '5' // '45'
-
-<b>undefined and null</b>
-undefined + 5 // NaN
-undefined + null // NaN
-null + 5 // 5
-
-<b>Object</b>
-{a: 5} + 6; // 6, as equals to some block and a <code>+6</code> 
-// a number with an unary operator <code>+</code>
-({a: 8} + 4); // '[object Object]4'
-[4] + 3; // '43'
-[3] + [4]; // '44'
-
-<b>BigInt</b>
-2n + 5; // Error, bigint will not be converted
-Number(2n) + 5; // 7
-4n + '5'; // '45
-
-<b>String</b>
-'abc' + 'cde' = 'abccde'
-
-</pre>                            
-                    
-                    `,
-                    Operator: '<code>+</code>',
-                    Name: `Addition`,
-                    Usage: 'g() + f()',
-                    Description: `
-                    In case of two number type operands returns a sum in case of two string operands returnes a 
-                    concatenated string. In case of other types performes a conversion. See code section for more 
-                    (in case of a non mobile device click this row)
-                    `
-                },
-
-
-
-
-
-                {
-                    [Symbol('title')]:'-',
-                    [Symbol('code')]:`                 
-<pre>
-5 - 6.5; // -1.5
-'5' - '4'// 1
-'5' - true; // 4
-4 - null; // 4
-
-'2n' - 1 // NaN conversion not possible
-3 - 's' // NaN conversion not possible
-undefined - null; // NaN, conversion not possible
-
-2n - 1; // Error, BigInt will not be converted by default
-Number(2n) - 1; // 1
-
-</pre>                            
-                    
-                    `,
-                    Operator: '<code>-</code>',
-                    Name: `Substraction`,
-                    Usage: 'g() - f()',
-                    Description: `
-                    Evaluates operands and returns their difference. There will be an attempt to 
-                    convert operands to a number if they are not a number and conversion is 
-                    possible
-                    `
-                },
-
-
-
-                {
-                    [Symbol('title')]:'*',
-                    [Symbol('code')]:`                 
-<pre>
-5 * 6.5; // 32.5
-'5' * '4'// 20
-'5' * true; // 5
-4 ) null; // 4
-
-'2n' * 1 // NaN conversion not possible
-3 * 's' // NaN conversion not possible
-undefined * null; // NaN, conversion not possible
-
-2n * 1; // Error, BigInt will not be converted by default
-Number(2n) * 1; // 2
-
-</pre>                            
-                    
-                    `,
-                    Operator: '<code>*</code>',
-                    Name: `Multiplication`,
-                    Usage: 'g() * f()',
-                    Description: `
-                    Evaluates operands and returns their multiplicated value. There will be an attempt to 
-                    convert operands to a number if they are not a number and conversion is 
-                    possible. 
-                    `
-                },
-
-
-
-
-
-                {
-                    [Symbol('title')]:'/',
-                    [Symbol('code')]:`                 
-<pre>
-5 / 6.5; // 0.7692307692307693
-'5' / '4'// 1.25
-'5' / true; // 5
-4 / null; // Infinity
-4 / Infinity; // 0
-4 / -Infinity; // -0
-
-'2n' / 1 // NaN conversion not possible
-3 / 's' // NaN conversion not possible
-undefined / null; // NaN, conversion not possible
-
-2n / 1; // Error, BigInt will not be converted by default
-Number(2n) / 1; // 2
-
-</pre>                            
-                    
-                    `,
-                    Operator: '<code>/</code>',
-                    Name: `Division`,
-                    Usage: 'g() / f()',
-                    Description: `
-                    Evaluates operands and returns their divided value. There will be an attempt to 
-                    convert operands to a number if they are not a number and conversion is 
-                    possible
-                    `
-                },
-
-
-
-
-
-
-
-
-
-            ]
-        },
-
-
-
-
-
-
-
-
-
-                {
-                    elementType:'Headline-3',
-                    content:'Assignment operators'
-                },
-                {
-                    elementType: 'Paragraph',
-                    content: `
-                        Assignment operators assigns its right operand value to its left operand value.
-                    `
-                },
-
-                {
-                    elementType: 'ConditionalArray',
-                    headItems: [
-                        'Operator','Name','Usage','Description'
-                    ],
-                    content: [
-                        {
-                            [Symbol('title')]:'=',
-                            [Symbol('code')]:`
-                            
-<pre>
-let sum = function(acc, element){
-    return acc + element;
-};
-let a = 0; let b = 1; let c = 2;
-let d = 3, e = 4; // d is 3, e is 4;
-e = a + b + [c, d, e].reduce(sum, 0);
-console.log(e);// 10;
-
-e = a = b = c = d;
-//e, a, b, c and d, all are equal to 3, as 'd' was 3
-//when assigned
-
-Chain assignments evaluate from left to right, but then they are assigned from
-right to left:
-a = (b = (c = (f() )))
-However chaining assignments is discouraged, aspecialy with <code>const</code>
-<code>let</code>, <code>var</code> it creates strage behaviours:
-let a = b = c = 5;
-will declare only a, and the rest (c, b) will:
-1) Not be declared at all and cause an error,
-2) May silently be declared as global variables.
-
-</pre>                            
-                            
-                            `,
-                            Operator: '<code>=</code>',
-                            Name: `Assignment`,
-                            Usage: 'a = f()',
-                            Description: 'Assigns the right evaluated expression to its left operand,'
-                        },
-
-
-
-
-                        {
-                            [Symbol('title')]:'+=',
-                            [Symbol('code')]:`          
-<pre>
-let a = 1;
-let b = 2
-let c = '4';
-let d = 'word';
-let e = 5;
-
-a += b; console.log(a); // 3;
-a += c; console.log(a); 
-// '34', because c was a string a gets converted to a string,
-// and the result is a string;
-
-d += a; // d is word34, as a was 34, and addition assignment 
-//concatenates strings,
-b =+ a; // b is now 34, as there is no =+
-//operator in JS, so '=' is an assignment, and '+ a'
-// converts a (that was a string of 34) back to the number, so
-// b is a number of 34
-
-
-</pre>                                   
-                            `,
-                            Operator: '<code>+=</code>',
-                            Name: `Addition assignment`,
-                            Usage: 'a += f()',
-                            Description: `Adds or concatenates the evaluated expression on the right to the left operand
-                            and assignes the result to the operand on the left,`
-                        },
-
-
-
-
-
-                        {
-                            [Symbol('title')]:'-=',
-                            [Symbol('code')]:`          
-<pre>
-let a = 'Hello World';
-a -= 'World'; // NaN, no string substraction
-let b = 10;
-b -= 1; // 9;
-</pre>                                   
-                            `,
-                            Operator: '<code>-=</code>',
-                            Name: `Substraction assignment`,
-                            Usage: 'a -= f()',
-                            Description: `
-                            <code>a = a - f()</code> 
-                            Substracts the evaluated expression on the right from the left operand, and assignes the
-                            reusult to the operand on the left,
-                            `
-                        },
-
-
-
-
-                        {
-                            [Symbol('title')]:'*=',
-                            [Symbol('code')]:`          
-<pre>
-let a = 'Hello World';
-a *= 'World'; // NaN, no string multiplication
-let b = 10;
-b *= 7; // 70;
-</pre>                                   
-                            `,
-                            Operator: '<code>*=</code>',
-                            Name: `Multiplication assignment`,
-                            Usage: 'a *= f()',
-                            Description: `
-                            <code>a = a * f()</code> 
-                            Multiplicates the evaluated expression on the right with the left operand and assignes to the 
-                            variable on the left,
-                            `
-                        },
-
-
-
-                        {
-                            [Symbol('title')]:'/=',
-                            [Symbol('code')]:`          
-<pre>
-let a = 'Hello World';
-a /= 2; // NaN, no string division
-let b, c, d, e;
-b = c = d = e = 70;
-b /= 7; // 10;
-c /= -0; // -Infinity;
-d /= +0; // +Infinity;
-</pre>                                   
-                            `,
-                            Operator: '<code>/=</code>',
-                            Name: `Division assignment`,
-                            Usage: 'a /= f()',
-                            Description: `
-                            <code>a = a / f()</code> 
-                            Divides the left operand by the evaluated expression on the rigth, and assigns the result to the
-                            operand on the left,
-                            `
-                        },
-
-
-                        {
-                            [Symbol('title')]:'%=',
-                            [Symbol('code')]:`          
-<pre>
-let a = 'Hello World';
-a %= 2; // NaN, no string division
-let b, c, d, e;
-b = c = d = e = 10;
-b /= 7; // 3;
-c /= -0; // NaN;
-d /= 0; // NaN;
-</pre>                                   
-                            `,
-                            Operator: '<code>%=</code>',
-                            Name: `Remainder assignment`,
-                            Usage: 'a %= f()',
-                            Description: `
-                            <code>a = a % f()</code> 
-                            Divides the left operand by the evaluated expression on the rigth, and assigns the remainder to the
-                            operand on the left,
-                            `
-                        },
-
-
-
-                        {
-                            [Symbol('title')]:'**=',
-                            [Symbol('code')]:`          
-<pre>
-let a = 'Hello World';
-a **= 2; // NaN, no raising a string to any power
-let b, c, d;
-b = c = 5;
-d = Infinity;
-b **= 3; // 125;
-b **= +0; // 1;
-d **= -0; // 1;
-b **= Infinity; // 1 ** Infinity == NaN
-c **= Infinity; // Infinity;
-d **= Infinity; // Infinity;
-</pre>                                   
-                            `,
-                            Operator: '<code>**=</code>',
-                            Name: `Exponention assignment`,
-                            Usage: 'a **= f()',
-                            Description: `
-                            <code>a = a ** f()</code> 
-                            Raises the value of a variable to the power of the right operand and assigns the result to the
-                            left operand
-                            `
-                        },
-
-
-
-
-                        {
-                            [Symbol('title')]:'<<=',
-                            [Symbol('code')]:`          
-<pre>
-let a = 5;
-let str = 'string';
-let inf = Infinity;
-let boolT = true;
-let boolF = false;
-a << =2; //20;
-a.toString(20); //10100
-str <<= 4; // 0;
-inf <<= 3; // 0;
-boolT <<= 3; // 8
-boolT.toString(2); // 1000
-boolF <<= 5; // 0;
-
-let m = -5;
-m <<= 3; // -40
-
-In general:
-n << m is n ** (2*m)
-</pre>                                   
-                            `,
-                            Operator: '<code><<=</code>',
-                            Name: `Right shift assignment`,
-                            Usage: 'a <<= f()',
-                            Description: `
-                            <code>a = a <<= f()</code> 
-                            Moves the specified amount of bits to the left and asigns the result to the left operand;
-                            `
-                        },
-
-
-
-                        {
-                            [Symbol('title')]:'>>=',
-                            [Symbol('code')]:`          
-<pre>
-let a, b, c;
-a = b = c  = 100;
-let str = 'string';
-let inf = Infinity;
-let boolT = true;
-let boolF = false;
-a >>= 1; // 50;
-b >>= 2; // 25;
-c >>= 3; // 12;
-str >>= 4; // 0;
-inf >>= 3; // 0;
-boolT >>= 0; // 1;
-boolT >>= 1; // 0;
-boolF >>= 5; // 0;
-
-let m = -100;
-m >>= 1;//-50
-
-In general:
-n >> m is floor(n ** (0.5 * m));
-</pre>                                   
-                            `,
-                            Operator: '<code>>>=</code>',
-                            Name: `Left shift assignment`,
-                            Usage: 'a >>= f()',
-                            Description: `
-                            <code>a = a >>= f()</code> 
-                            Moves the specified amount of bits to the right and asigns the result to the left operand;
-                            If a negative number is processed, then the literal negative value is taken
-                            `
-                        },
-
-
-                        {
-                            [Symbol('title')]:'>>>=',
-                            [Symbol('code')]:`          
-        <pre>
-See representation of negative numbers section for
-more details.
-let positive =  100;
-let negative = -100;
-positive >>>= 1; // 50;
-negative >>>= 1; // 2147483598
--100 >> 1; // -50
-
-        </pre>                                   
-                            `,
-                            Operator: '<code>>>>=</code>',
-                            Name: `Unsigned right shift assignment`,
-                            Usage: 'a >>>= f()',
-                            Description: `
-                            <code>a = a >>>= f()</code> 
-                            Moves the specified amount of bits to the right and asigns the result to the left operand;
-                            `
-                        },
-
-
-                        {
-                            [Symbol('title')]:'&=',
-                            [Symbol('code')]:`          
-<pre>
-let a = 5; // 101;
-let b = 2; // 010;
-a &= b; // 0 & 1 is 0, so 000 is the result;
-console.log(a); // 0
-
-let c = 5;
-c &= 3; // 101 and 011, common bit is 001,
-console.log(c); // 1
-
-let d = -5; // 11111111111111111111111111111011
-// this is the negative number representatnion:
-// (a >>> 0).toString(2);
-d &= 5; // 00...00101
-//common bit of 1011 and 0101 is 0001
-console.log(d); // 1
-
-
-</pre>                                   
-                            `,
-                            Operator: '<code>&=</code>',
-                            Name: `Bitwise AND assignment`,
-                            Usage: 'a &= f()',
-                            Description: `
-                            <code>a = a &= f()</code> 
-                            Takes a binary representation of both operands, does a bitwise AND operation on them 
-                            and assigns the result to the left operator,
-                            `
-                        },
-
-
-                        {
-                            [Symbol('title')]:'^=',
-                            [Symbol('code')]:`          
-<pre>
-0 ^ 0 == 0;
-1 ^ 1 == 0;
-0 ^ 1 == 0;
-1 ^ 0 == 0;
-let a = 5;
-a ^= 3; // 110 is 6
-let b = -5;
-b ^= 3; 
-// 11111111111111111111111111111011 ^
-// 00000000000000000000000000000101 =
-// 11111111111111111111111111111000 
-// and that is -8
-
-</pre>                                   
-                            `,
-                            Operator: '<code>^=</code>',
-                            Name: `Bitwise XOR assignment`,
-                            Usage: 'a ^= f()',
-                            Description: `
-                            <code>a = a ^= f()</code> 
-                            Takes a binary representation of both operands, does a bitwise XOR on them and assigns the 
-                            result to the left operator,
-                            `
-                        },
-
-
-
-
-                        {
-                            [Symbol('title')]:'|=',
-                            [Symbol('code')]:`          
-<pre>
-let a = 5;
-a |= 3; // 101 | 011 is 111 so 7
-let b = -8;
-b |= 8; // -8
-// -8 is  11111111111111111111111111111000
-// +8 is  00000000000000000000000000001000
-// result 11111111111111111111111111111000
-
-</pre>                                   
-                            `,
-                            Operator: '<code>|=</code>',
-                            Name: `Bitwise OR assignment`,
-                            Usage: 'a |= f()',
-                            Description: `
-                            <code>a = a |= f()</code> 
-                            Takes a binary representation of both operands, does a bitwise OR on them and assigns the 
-                            result to the left operator,
-                            `
-                        },
-
-                        {
-                            [Symbol('title')]:'||=',
-                            [Symbol('code')]:`          
-<pre>
-false || false === false;
-true || false === true;
-false || true === true;
-true || true === true;
-
-so:
-let a = false;
-a ||= false; //false
-a ||= true; //true
-a ||= false; // true, as it was already true in previous step;
-
-let b = 5;
-b ||= true; //5
-b ||= false; //5
-
-let c = 0;
-c ||= true; // 1
-
-5 || 7 === 5;
-7 || 5 === 7;
-null || 'something' === 'something'
-
-false || null === null
-false || undefined === undefined
-null || false === false;
-// it is important what will be the right side operand
-
-NOTE:
-let notChanged = 0;
-let f = () => {
-    notChanged = 5;
-    return true;
-}
-let d = true;
-d ||= f();
-// notChanged is still 0,
-// as there was no need to evaluate it
-
-</pre>                                   
-                            `,
-                            Operator: '<code>||=</code>',
-                            Name: `Logical OR assignment`,
-                            Usage: 'a ||= f()',
-                            Description: `
-                            <code>x = y || f()</code> <code>x || (y = z)</code>, not <code> x = x || y</code>
-                            Evaluates the right side operand expression <strong>only if the left site does not 
-                            determine the result</strong> and assigns to x <strong>only if the right sight
-                            is evaluated and would change the left site operand value</strong>. However if the 
-                            right side operand is evaluated, the assertion will take place 
-                            <code>null || undefined === undefined</code> 
-                            `
-                        },
-
-                        {
-                            [Symbol('title')]:'&&=',
-                            [Symbol('code')]:`          
-<pre>
-false && false === false;
-true  && false === false;
-false && true === false;
-true  && true === true;
-
-so:
-let a = false;
-a &&= false; //false
-a &&= true; //false
-a &&= false; // No possibility to make true out of this with &&
-let b = true;
-b &&= true; //true
-b &&= false; //false
-b &&= true; //false, as was made false in previous line
-
-let b = 6;
-b &&= true; // true
-b &&= false; // false
-
-let c = true;
-c &&= 5; // 5
-c &&= false; 
-c &&= 5; //false
-
-5 && 7 === 7;
-7 && 5 === 5;
-null && 'something' === null;
-
-NOTE:
-let notChanged = 0;
-let f = () => {
-    notChanged = 5;
-    return true;
-}
-let d = false;
-d &&= f();
-// notChanged is still 0,
-// as there was no need to evaluate it
-
-</pre>                                   
-                            `,
-                            Operator: '<code>&&=</code>',
-                            Name: `Logical AND assignment`,
-                            Usage: 'a &&= f()',
-                            Description: `
-                            <code>x = y && f()</code> <code>x && (y = z)</code>, not <code> x = x && y</code>
-                            Evaluates the right side operand expression <strong>only if the left side does not 
-                            determine the result</strong> and assigns to x the x AND f() result <strong>only if
-                            necessarry</strong>
-                            `
-                        },
-
-
-
-                        {
-                            [Symbol('title')]:'??=',
-                            [Symbol('code')]:`          
-<pre>
-let a = null;
-let b = undefined;
-a ??= undefined; 
-// a is undefined
-b ??= null;
-// b is null;
-a ??= 5; // a is 5
-b ??= 'val'; // b is 'val'
-a ?? 7; // a is 5, as it was not undefined or null
-
-NOTE:
-let notChanged = 0;
-let f = () => {
-    notChanged = 5;
-    return true;
-}
-let d = '';
-d ??= f();
-// notChanged is still 0,
-// as there was no need to evaluate it, d was != undefined or null
-
-</pre>                                   
-                            `,
-                            Operator: '<code>??=</code>',
-                            Name: `Logical nullish assignment`,
-                            Usage: 'a ??= f()',
-                            Description: `
-                            Assigns a right hand expression only if left hand operand is <code>null</code> or
-                            <code>undefined</code>. If left hand operand is not null or undefined the right hand
-                            expression is not evaluated at all.
-                            `
-                        },
-
-
-
-            ]
-            },
 
             ]
         },
